pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "omni"
version = "0.0.1a1"
requires-python = ">=3.10"
description = "DreamLLM: Synergistic Multimodal Comprehension and Creation."
readme = "README.md"
license = { file = "LICENSE" }
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    "av",
    "tqdm",
    "fire",
    "ipdb",
    "numpy",
    "wandb",
    "isort",
    "pyzmq",
    "black",
    "boto3",
    "scipy",
    "awscli",
    "psutil",
    "dacite",
    "joblib",
    "loguru",
    "openai",
    "megfile",
    "pandas",
    "imageio",
    "isodate",
    "uvicorn",
    "fastapi",
    "pendulum",
    "requests",
    "openpyxl",
    "beartype",
    "tabulate",
    "influxdb",
    "colorama",
    "jsonlines",
    "shortuuid",
    "netifaces",
    "omegaconf",
    "smart_open",
    "ema_pytorch",
    "Levenshtein",
    "pycocotools",
    "pyinstrument",
    "editdistance",
    "nvidia-ml-py3",
    "opencv-python",
    "markdown2[all]",
    "pytorch-warmup",
    "gateloop_transformer",
    "peft==0.3.0",
    "redis>=3.4.0",
    "timm==0.6.13",
    "httpx==0.24.0",
    "einops==0.7.0",
    "gradio==3.35.2",
    "Flask>=0.10.1",
    "kornia==0.7.0",
    "msgpack>=1.0.0",
    "deepspeed==0.12.6",
    "datasets==2.16.1",
    "netifaces>=0.10.6",
    "diffusers==0.24.0",
    "hydra-core==1.3.2",
    "tokenizers==0.15.0",
    "urllib3<2,>=1.21.1",
    "pycocoevalcap==1.2",
    "einops-exts==0.0.4",
    "bitsandbytes==0.41.0",
    "accelerate==0.23.0",
    "transformers==4.35.2",
    "sentencepiece==0.1.99",
    "huggingface-hub==0.19.4",
    "websocket-client==0.43.0",
]
[project.scripts]
smartrun = "omni:dispatch_torchrun"

[project.urls]
# Homepage = "https://example.com"
# Documentation = "https://readthedocs.org"
Repository = "https://github.com/BaseLLVM/omni"
"Bug Tracker" = "https://github.com/BaseLLVM/omni/issues"
Changelog = "https://github.com/BaseLLVM/omni/blob/main/CHANGELOG.md"

[tool.setuptools]
packages = ["omni"]

[tool.black]
line-length = 128
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 128
